Insteon: Add User Option to Enable PLM Restart; Prevent Crashes
This adds a new INI variable

Insteon_PLM_reconnect_count

By default this is set to 99, to disable any PLM reconnect attempts.  If you have PLM disconnect issues, try setting it to 2 or 3.  The setting waits for the number of missing ACK responses from the PLM before trying to reconnect.

This also fixes an issue where trying to close an undefined port would completely crash MH, now it just results in a warning, but MisterHouse continues.