Fix Link Generation in Items and Modules Documenation Pages
I don't know when or how this broke, but I made some changes to update_docs some months back.

I noticed that on my local copy, the links on the Items and Modules pages were non-existant.  This change fixes that.

I also made a change to the pod.css so that CODE tags are mono-spaced formatted.