Nest: Move Write Data to a Process Item; Check Write Response
This will prevent the write action from blocking if there is a delay connecting to the Nest server.  The write process is a little complicated but should be very robust, it can handle queuing multiple write commands at once.

Added a simple check of the write response, will now print an error message if the response is not OK.

Fixes krkeegan/misterhouse#27
Fixes krkeegan/misterhouse#26