HTTP_Server: Properly Handle Quoted Strings in a Sub Request
The prior code could only handle quoted strings if the entire argument was a single quote.  This change allows for calls like:

```
/sub?json(groups=Living_Room,"fields=text|type|state|states|label,not_recursive")
```

Note that the second argument is a string joined together by a comma.  The prior code clobbered this into two separate arguments.

The only potential hiccup, is the new code relies on Text::ParseWords.  This is listed as a standard package, but sometime you never know with perl.  I am slightly concerned because the http server is a base componet of MisterHouse.