IA7/JSON: Change JSON Structure; Enable Updating in IA7
Added request and option to returned json to make recreating the request easier.

Added millisecond time to JSON response to allow for long polling.

Enable long polling in JSON.  If long poll is requested, the routine will return nothing if only an empty JSON would be returned.  In addition, the time parameter is used to skip objects which have not changed since time.  This used the idle_time value.