Generic_Item/Group: Add Function to Store Sort Order of Members
Function simply stores an ordered list of names.  The ordered list can be retreived for sorting the members.

While this is added to Generic_Item, it is likely only available on Group items because it seems that the general practice was not to call Generic_Item->new on sub items that inherit Generic_Item.  This seems wrong to me.