Insteon: Change the Nomeclature of ALDB Health
It was confusing to list both links and link tables as "out-of-sync."

At least for link tables, the likely command would be "sync ...." but in fact it is "scan ...".

Now, link tables can be unknown|empty|changed|unchanged.

This fits with the prior voice command "scan changed link tables."

Hopefully this will lead to less confusion.