Further work on the mqtt module
I've made a number of changes to allow more than one mqtt instance. This is
meant to provide support for multiple hosts and ports pairs. I noticed that
using the same host:port would cause the second request to be disconnected. So I
reused the first host:port instance (I really only wanted the socket). So I
reused the instance and issued a new subscription. This has man side effects
that need to be worked out.

There are a number of other fix "me"s and todos (listed in the source).
Reconnect handleing among the top issues.

See Usage and Example in the source file for changes to instanctiation.