Add Travis CI regression test support
This pull request adds the required files to run a basic regression test on the MisterHouse code using Travis.

All control is in the `.travis.yml` file in the root of the repository.

Every time somebody commits code, a new build is triggered on Travis.

The test code that is executed is:
```
mh -code_dir code/test
```

With the current code that is present in that folder, MisterHouse is started, runs for one minute and then exits.

Results are here: https://travis-ci.org/hollie/misterhouse/builds

Rationale:
 * to offer a framework that allows automatic code verification for new pull requests
 * to speed up acceptance of new pull requests that pass the tests

Note:
 * I had to change the exit code of MisterHouse to '0' for an intended exit (so that the test can stop at a certain point in time). That exit code used to be 1, and I changed both the exit code and the loop code that checks the exit code. We need to check if there are other uses of the exit code '1' that could be broken by this change
 * More test code needs to be added to the `code/test/` folder. E.g. performing a basic check if the web interface is up, adding specific test code for modules, ...

So: this is a start, I hope we can expand it to cover larger parts of the MisterHouse code.