Add support for the AoG Smart Home Provider
Add support for the Actions on Google Smart Home Provider
so the Google Assistant (on the Google Home and other
devices like smartphones) can interact with MisterHouse.

An Actions on Google Smart Home Provider implementation requires four
things:

1. An OAuth server to the user can link a MisterHouse account to the
Actions on Google Smart Home Provider from the Google Home app.

2. A handler for a "sync" request from the Google Assistant. This tells
the Google Assistant what MisterHouse devices are available for control.

3. A handler for a "query" request from the Google Assistant. The
Assistant sends "query" requests when it needs to know the state of a
device.

4. A handler for an "execute" request from the Google Assistant. The
Assistant sends "execute" requests when the user requests an action to
be performed, like turning on or off a light, dimming or brightening a
light, etc.

See https://developers.google.com/actions/smarthome/ for details.

This initial commit adds supports to MisterHouse for all the four pieces
described above.

The devices that are to be exposed to the Actions on Google Smart Home
Provider must be specified via AOGSMARTHOME_ITEMS and AOGSMARTHOME_ITEM
items (which are new) in a .mht file (or via user code).

This implementation adds a hook to the MisterHouse HTTP server
(lib/http_server.pl) to process two specific HTTP requests from the
Actions on Google Smart Home Provider, one for the OAuth flow, and the
other for the sync, query, and execute commands. The code that handles
these two HTTP requests lives in the new file lib/http_server_aog.pl.
Everything else lives in the new file lib/AoGSmartHome_Items.pm.

The commit also makes some small cleanups to lib/http_server.pl that are
not related to the Actions on Google Smart Home Provider implementation.
Specifically:

* Use the m// and s// operators with delimiters other than '/' so heavy
escaping of '/' is not necessary, e.g. m%/path/to/file% instead of
m/\/path\/to\/file/. This makes the regular expressions easier to read,
especially when the code handles HTTP requests that contain paths.

* Match $Http{'Content-Type'} against "application/json" using a regular
expression as "if (lc($Http{'Content-Type'}) eq lc('application/json')
)" breaks when Content-Type is "application/json;charset=UTF-8", which
the Actions on Google Smart Home Provider is sending.