Do not try to get device status before responding
Do not try to get the status of a device before sending the response to
the EXECUTE command -- it does not work because we (apparently) have to
run the main loop to get device status update, so this does not seem
possible to do from the HTTP server. Glad to be proven wrong, but
for now just removing this code that does not work.