Unconditionally send "on" state in QUERY response
Send an "on" state unconditionally for all devices of type "Light".
Before we were not sending the "on" state for devices that are dimmable,
which is incorrect because a dimmable light is either on, or off. This
caused the Google Assistant command "what lights are on?" to fail for
dimmable lights.

A response to a QUERY command now looks like this:

{
 "requestId": "18352987651883269961",
 "payload": {
  "devices": {
   "11": {
    "on": false,
    "online": true
   },
   "6": {
    "on": false,
    "brightness": 0,
    "online": true
   },
[...]