CBus->CGate Socket checks
Test the CGate Talker and Monitor sockets for activity (i.e. health) on each loop. If one is inactive, try to reconnect it (but not on every loop, to minimise log clutter.) Reduce the frequency of CBus network sync queries when the network is not in sync, from every loop, to every 100 loops, to allow low power hardware to catch up. This makes the CBus module more robust, accomodates connections to CGate going away, and better accomodates startup where the CBUs network has not yet attained sync.