raZberry: Login again after session expires
If the session to z-way expires, a 401-message is receive after sending a
command. This change clears all existing cookies and starts a new login
after 401 was seen.
The internal command queue was change to store only the command url
instead of the complete commandline. The commandline is now created
before each request to use always the latest session cookie.