Always show error message on failed ajax requests
If misterhouse is run behind web proxy (tested with nginx), ajax
requests may be answered with an html error page if misterhouse is not
running. In this case no valid JSON is returned. This fix simply catches
the JSON parse exception and adds http status + text to the system
message. Before, the message was never shown because the exception was not
handled.